gwhf: Properly handle keep-alive connection
Currently, the app always treats all connections as keep alive until the
other side closes the socket. Make sure the app properly terminates the
connection based on the "Connection" header and the HTTP version.

For an HTTP/1.1 request without a "Connection" header, treat it as
keep-alive. For an HTTP/1.0 request without a "Connection" header, treat
it as close. If "Connection: keep-alive" is explicitly provided, treat
it as keep-alive for both versions.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>