gwhf: Keep track of file changes in the caching system
The current file caching system caches the file size as well. Now,
ensure the cached file is appropriately adjusted to the new file size
when the file size changes. With this change, if the file is changed but
the file size is still the same, it will result in the new file in the
response body, so there is no need to worry about such a case; it's
always valid to match the file size. The main goal of this change is to
make the response body up-to-date.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>