gwhf: ev/epoll: Avoid too long loop in the send buffer
When a local client downloads a large file, the loop in the send buffer
can take too long to finish or reach EAGAIN. It causes other clients to
get ignored for a long time.

Stop looping when it has reached 16 iterations; try to poll for other
events before continuing the send.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>